# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vLLM Verifier",
    "description": "Jev-compatible decisions with uncalibrated model estimates.",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Healthz Get"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/readyz": {
      "get": {
        "summary": "Readiness",
        "operationId": "readiness_readyz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/models": {
      "get": {
        "summary": "Models",
        "operationId": "models_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Models V1 Models Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/systemone": {
      "post": {
        "summary": "System One",
        "operationId": "system_one_v1_systemone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemOneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/vision/systemone": {
      "post": {
        "summary": "Vision One",
        "operationId": "vision_one_v1_vision_systemone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Choice": {
        "properties": {
          "type": {
            "type": "string",
            "const": "choice",
            "title": "Type"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "criteria": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object"
                },
                {
                  "items": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ]
            },
            "propertyNames": {
              "maxLength": 256,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 255,
            "minProperties": 1,
            "title": "Criteria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "criteria"
        ],
        "title": "Choice"
      },
      "ChoiceAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "choice",
            "title": "Type",
            "default": "choice"
          },
          "choice": {
            "type": "string",
            "title": "Choice"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0
            },
            "type": "object",
            "title": "Probabilities"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "choice",
          "probabilities",
          "confidence"
        ],
        "title": "ChoiceAnswer"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JsonValue": {},
      "Noul": {
        "properties": {
          "type": {
            "type": "string",
            "const": "noul",
            "title": "Type"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "criteria": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NoulCriteria"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Noul"
      },
      "NoulAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "noul",
            "title": "Type",
            "default": "noul"
          },
          "noul": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Noul"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "noul"
        ],
        "title": "NoulAnswer"
      },
      "NoulCriteria": {
        "properties": {
          "true": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "True"
          },
          "false": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "False"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NoulCriteria"
      },
      "Score": {
        "properties": {
          "type": {
            "type": "string",
            "const": "score",
            "title": "Type"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "criteria": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object"
                },
                {
                  "items": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Criteria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "criteria"
        ],
        "title": "Score"
      },
      "ScoreAnswer": {
        "properties": {
          "type": {
            "type": "string",
            "const": "score",
            "title": "Type",
            "default": "score"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0
            },
            "type": "object",
            "title": "Probabilities"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "legend": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object"
                },
                {
                  "items": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Legend"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "score",
          "probabilities",
          "confidence",
          "legend"
        ],
        "title": "ScoreAnswer"
      },
      "SystemOneRequest": {
        "properties": {
          "model": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Model"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              }
            ],
            "title": "State"
          },
          "questions": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Choice"
                },
                {
                  "$ref": "#/components/schemas/Score"
                },
                {
                  "$ref": "#/components/schemas/Noul"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/Choice",
                  "noul": "#/components/schemas/Noul",
                  "score": "#/components/schemas/Score"
                }
              }
            },
            "propertyNames": {
              "maxLength": 256,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 64,
            "minProperties": 1,
            "title": "Questions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model",
          "state",
          "questions"
        ],
        "title": "SystemOneRequest"
      },
      "SystemOneResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "answers": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceAnswer"
                },
                {
                  "$ref": "#/components/schemas/ScoreAnswer"
                },
                {
                  "$ref": "#/components/schemas/NoulAnswer"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/ChoiceAnswer",
                  "noul": "#/components/schemas/NoulAnswer",
                  "score": "#/components/schemas/ScoreAnswer"
                }
              }
            },
            "type": "object",
            "title": "Answers"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model",
          "answers",
          "usage"
        ],
        "title": "SystemOneResponse"
      },
      "Usage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Tokens",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Usage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisionRequest": {
        "properties": {
          "model": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Model"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "array"
              }
            ],
            "title": "State"
          },
          "questions": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Choice"
                },
                {
                  "$ref": "#/components/schemas/Score"
                },
                {
                  "$ref": "#/components/schemas/Noul"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/Choice",
                  "noul": "#/components/schemas/Noul",
                  "score": "#/components/schemas/Score"
                }
              }
            },
            "propertyNames": {
              "maxLength": 256,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 64,
            "minProperties": 1,
            "title": "Questions"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Images"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model",
          "state",
          "questions",
          "images"
        ],
        "title": "VisionRequest"
      },
      "VisionResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "answers": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceAnswer"
                },
                {
                  "$ref": "#/components/schemas/ScoreAnswer"
                },
                {
                  "$ref": "#/components/schemas/NoulAnswer"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/ChoiceAnswer",
                  "noul": "#/components/schemas/NoulAnswer",
                  "score": "#/components/schemas/ScoreAnswer"
                }
              }
            },
            "type": "object",
            "title": "Answers"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "observation": {
            "type": "string",
            "title": "Observation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model",
          "answers",
          "usage",
          "observation"
        ],
        "title": "VisionResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}